♻️(front) only export connected VideoPlayer component
We want to test if it's possible to test a component if we only export
it's connected version to redux. And we test it with
react-testing-library.
