+++ src/frontend/components/AppRoutes/AppRoutes.tsx
@@ -13,8 +13,8 @@ import { REDIRECT_ON_LOAD_ROUTE } from '../RedirectOnLoad/route';
 import { RedirectOnLoadConnected } from '../RedirectOnLoadConnected/RedirectOnLoadConnected';
 import { UPLOAD_FORM_ROUTE } from '../UploadForm/route';
 import { UploadFormConnected } from '../UploadFormConnected/UploadFormConnected';
+import { VideoPlayer } from '../VideoPlayer';
 import { VIDEO_PLAYER_ROUTE } from '../VideoPlayer/route';
-import { VideoPlayerConnected } from '../VideoPlayerConnected/VideoPlayerConnected';
 
 interface AppRoutesProps {
   context: RootState<appState>['context'];
@@ -29,7 +29,7 @@ export const AppRoutes = ({ context }: AppRoutesProps) => {
           path={VIDEO_PLAYER_ROUTE()}
           render={() => (
             <InstructorWrapperConnected>
-              <VideoPlayerConnected
+              <VideoPlayer
                 video={context.ltiResourceVideo}
                 createPlayer={createPlayer}
               />
